pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speech2dialogue"
version = "1.0.0"
description = "完全离线的语音/视频转对话脚本生成器"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Speech2Dialogue Team"}
]
keywords = ["speech-to-text", "whisper", "diarization", "asr", "offline"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.0.0",
    "torchaudio",
    "whisperx>=4.0.0",
    "faster-whisper",
    "pyannote.audio",
    "modelscope",
    "soundfile",
    "pydub",
    "onnxruntime",
    "noisereduce",
    "librosa",
    "speechbrain",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
]

[project.scripts]
speech2dialogue = "speech2dialogue.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/speech2dialogue"
Repository = "https://github.com/yourusername/speech2dialogue"
Issues = "https://github.com/yourusername/speech2dialogue/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["speech2dialogue*"]
exclude = ["tests*", "outputs*", "models*"]

[tool.setuptools.package-data]
speech2dialogue = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
